crawler/requirement.txt
scrapling
redis
psycopg2-binary
beautifulsoup4
lxml